refactor: Move data classes into structures folder
