+++ source/hitomi.ts
@@ -4,7 +4,7 @@ import { TagManager } from './managers/tag';
 import { DEFAULT_HEADERS, RESOURCE_DOMAIN, MAXIMUM_AGE_PROPERTIES } from './internal/constants';
 import { defineProperties, capitalize } from './internal/functions';
 import { IndexProvider, Provider } from './internal/structures';
-import { HitomiError } from './error';
+import { HitomiError } from './structures/error';
 import { request, type RequestFunction, hash, type HashFunction, ResponseType, toString, RequestContext, OnRequestFunction } from '@platform';
 
 /**
